feat: Int배열의 모든 원소들을 더하는 Calculator 구현
- int배열의 모든 원소를 더한 값을 return하는 arrSum() 구현

+++ src/main/java/calculator/controller/CalculatorController.java
@@ -19,7 +19,7 @@ public class CalculatorController {
     public void calculate() {
         outputView.printPrompt();
         String input = inputView.strInput();
-        Long sum = calculatorService.calculateSumFromString(input);
+        int sum = calculatorService.calculateSumFromString(input);
     }
 
 }
